+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0+4
+* Add `deveoperId` and `accountId` when requesting `purchase` or `subscription` in `android`. Find out more in `requestPurchase` and `requestSubscription`.
+
+## 2.0.0+3
+* Correctly mock invoke method and return results [#94](https://github.com/dooboolab/flutter_inapp_purchase/pull/96)
+
 ## 2.0.0+2
 * Seperate long `example` code to `example` readme.